Refactored database update code
