+++ server/package.json
@@ -5,7 +5,7 @@
     "prestart": "npm run transpile",
     "start": "node ./dist/server.js",
     "transpile": "node_modules/.bin/babel src -d dist",
-    "update": "npm run transpile && node ./dist/utils/updateDatabase"
+    "update": "npm run transpile && node ./dist/db/update"
   },
   "dependencies": {
     "axios": "^0.19.2",
