+++ server/src/db/models.js
@@ -1,0 +1,13 @@
+import Aircraft from '../models/aircraft';
+import Airline from '../models/airline';
+import Airport from '../models/airport';
+import Country from '../models/country';
+import Region from '../models/region';
+
+export default {
+  aircraft: Aircraft,
+  airlines: Airline,
+  airports: Airport,
+  countries: Country,
+  regions: Region,
+};
