+++ server/src/models/airport.js
@@ -1,8 +1,9 @@
-import parse from 'csv-parse/lib/sync';
 import { model, Schema } from 'mongoose';
 
 import { CountrySchema } from './country';
 import { RegionSchema } from './region';
+
+import { parseOurAirportsData } from '../db/parseData';
 
 export const AirportSchema = new Schema({
   _id: String,
@@ -28,11 +29,7 @@ export const AirportSchema = new Schema({
 
 AirportSchema.static('dataUrl', 'https://ourairports.com/data/airports.csv');
 
-AirportSchema.static('parseData', data => {
-  return parse(data, { skip_empty_lines: true })
-    .slice(1)
-    .map(row => model.getUpdate(row));
-});
+AirportSchema.static('parseData', parseOurAirportsData);
 
 AirportSchema.static('getUpdate', row => ({
   _id: row[1],
